requirements.txt
alembic==1.0.7
amqp==2.4.0
aniso8601==4.1.0
asn1crypto==0.24.0
astroid==1.6.3
bugsnag==3.5.0
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudinary==1.15.0
coverage==4.5.2
cryptography==2.5
csvalidate==1.1.1
et-xmlfile==1.0.1
Faker==1.0.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Excel==0.0.7
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.1
Flask-RESTful==0.3.6
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flatten-dict==0.0.3.post1
gspread==3.1.0
httplib2==0.12.0
humps==0.2.2
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
lml==0.0.9
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.16.0
marshmallow-enum==1.4.1
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
numpy==1.16.1
oauth2client==4.1.3
openpyxl==2.5.14
pathlib2==2.3.3
pbr==5.1.2
pep8==1.7.1
pluggy==0.8.1
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
pyexcel-xlsxw==0.4.2
PyJWT==1.7.1
pylint==1.8.4
pytest==4.2.0
pytest-cov==2.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.4
python-http-client==3.1.0
pytz==2018.9
redis==2.10.6
requests==2.21.0
rsa==4.0
sendgrid==5.6.0
six==1.12.0
slackclient==1.3.0
SQLAlchemy==1.2.17
text-unidecode==1.2
texttable==1.6.0
typed-ast==1.1.1
urllib3==1.24.1
vine==1.2.0
WebOb==1.8.5
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
xlrd==1.2.0
XlsxWriter==0.9.3
xlwt==1.3.0
yapf==0.25.0